refactor : improving project packages
