--- app/src/main/java/com/neo/fbrules/main/domain/model/parse.kt
@@ -1,10 +1,0 @@
-package com.neo.fbrules.main.domain.model
-
-import com.neo.fbrules.main.data.model.DataCredential
-
-fun DomainCredential.toData(): DataCredential {
-    return DataCredential(
-        privateKey = privateKey,
-        databaseKey = databaseKey
-    )
-}
