+++ app/src/main/java/com/neo/fbrules/main/presenter/MainActivity.kt
@@ -17,7 +17,7 @@ import com.google.firebase.analytics.ktx.logEvent
 import com.google.firebase.ktx.Firebase
 import com.neo.fbrules.BuildConfig
 import com.neo.fbrules.R
-import com.neo.fbrules.main.domain.model.RulesEditor
+import com.neo.fbrules.main.presenter.contract.RulesEditor
 import com.neo.fbrules.main.presenter.adapter.NeoUtilsAppsAdapter
 import com.neo.fbrules.main.presenter.fragment.bottomSheet.ConfigBottomSheet
 import com.neo.fbrules.main.presenter.fragment.dialog.EncryptionDialog
